refactor refactor - TipoDocumento - Dto - ExpresionRegular
