+++ src/main/java/com/asisge/consultifybackend/autenticacion/aplicacion/dto/CambioContrasenaDto.java
@@ -1,6 +1,6 @@
 package com.asisge.consultifybackend.autenticacion.aplicacion.dto;
 
-import com.asisge.consultifybackend.usuarios.aplicacion.dto.Dto;
+import com.asisge.consultifybackend.utilidad.dominio.modelo.Dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
