+++ src/main/java/com/asisge/consultifybackend/autenticacion/infraestructura/controlador/handler/ApiRestExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.asisge.consultifybackend.autenticacion.infraestructura.controlador.handler;
 
 import com.asisge.consultifybackend.autenticacion.aplicacion.dto.ApiError;
-import com.asisge.consultifybackend.autenticacion.infraestructura.excepcion.ViolacionIntegridadException;
+import com.asisge.consultifybackend.utilidad.dominio.excepcion.ViolacionIntegridadException;
 import jakarta.validation.ConstraintViolationException;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -52,6 +52,13 @@ public class ApiRestExceptionHandler extends ResponseEntityExceptionHandler {
         ApiError error = new ApiError(HttpStatus.BAD_REQUEST.value(), ex.getMessage(), ex.getMessage());
         return handleExceptionInternal(ex, error, new HttpHeaders(), HttpStatus.BAD_REQUEST, request);
     }
+
+    @ExceptionHandler(value = {NullPointerException.class})
+    protected ResponseEntity<Object> handleNullPointerException(NullPointerException ex, WebRequest request) {
+        ApiError error = new ApiError(HttpStatus.INTERNAL_SERVER_ERROR.value(), ex.getMessage(), ex.getMessage());
+        return handleExceptionInternal(ex, error, new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR, request);
+    }
+
 
     @ExceptionHandler(value = {BadCredentialsException.class})
     protected ResponseEntity<Object> handleBadCredentialsException(BadCredentialsException ex, WebRequest request) {
