+++ src/main/java/com/asisge/consultifybackend/usuarios/aplicacion/dto/NuevoUsuarioAutenticadoDto.java
@@ -1,7 +1,8 @@
 package com.asisge.consultifybackend.usuarios.aplicacion.dto;
 
 import com.asisge.consultifybackend.usuarios.dominio.modelo.Rol;
-import com.asisge.consultifybackend.usuarios.dominio.modelo.TipoDocumento;
+import com.asisge.consultifybackend.utilidad.dominio.modelo.Dto;
+import com.asisge.consultifybackend.utilidad.dominio.modelo.TipoDocumento;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
