+++ src/main/java/com/asisge/consultifybackend/usuarios/aplicacion/dto/UsuarioBasicoDto.java
@@ -1,5 +1,6 @@
 package com.asisge.consultifybackend.usuarios.aplicacion.dto;
 
+import com.asisge.consultifybackend.utilidad.dominio.modelo.Dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
