+++ src/main/java/com/asisge/consultifybackend/utilidad/dominio/modelo/Dto.java
@@ -1,0 +1,6 @@
+package com.asisge.consultifybackend.utilidad.dominio.modelo;
+
+public interface Dto {
+
+    boolean validarDto();
+}
